🚚 Move remaining domain components out of components/ (RAL-1291)
- visibility.tsx -> features/poll/components/
- pro-badge.tsx, upgrade-button.tsx -> features/billing/components/
- logo.tsx -> features/branding/components/
- add-to-calendar-button.tsx -> features/calendars/components/
- user-provider.tsx, user-dropdown.tsx -> features/user/components/
- components/ now has zero @/features imports; the src/components/**
  entries in the biome.json migration allowlist are gone
