+++ apps/web/biome.json
@@ -120,12 +120,6 @@
     },
     {
       "includes": [
-        "src/components/add-to-calendar-button.tsx",
-        "src/components/logo.tsx",
-        "src/components/pro-badge.tsx",
-        "src/components/upgrade-button.tsx",
-        "src/components/user-provider.tsx",
-        "src/components/visibility.tsx",
         "src/features/billing/components/pay-wall-dialog.tsx",
         "src/features/poll/components/manage-poll.tsx",
         "src/lib/auth.ts",
