+++ apps/web/src/app/[locale]/(optional-space)/poll/[urlId]/guest-poll-alert.tsx
@@ -4,7 +4,7 @@ import { InfoIcon } from "lucide-react";
 
 import { LoginLink } from "@/components/login-link";
 import { RegisterLink } from "@/components/register-link";
-import { useUser } from "@/components/user-provider";
+import { useUser } from "@/features/user/components/user-provider";
 import { Trans } from "@/i18n/client";
 
 export const GuestPollAlert = () => {
