+++ apps/web/src/features/billing/components/pay-wall-dialog.tsx
@@ -29,7 +29,7 @@ import {
 } from "lucide-react";
 import React from "react";
 import { PageIcon } from "@/components/page-icons";
-import { UpgradeButton } from "@/components/upgrade-button";
+import { UpgradeButton } from "@/features/billing/components/upgrade-button";
 import type { SpaceTier } from "@/features/space/schema";
 import { spaceTierSchema } from "@/features/space/schema";
 import { Trans } from "@/i18n/client";
