+++ apps/web/src/features/poll/components/discussion.tsx
@@ -22,7 +22,6 @@ import { MoreHorizontalIcon, TrashIcon } from "lucide-react";
 import * as React from "react";
 import { Controller, useForm } from "react-hook-form";
 import { OptimizedAvatarImage } from "@/components/optimized-avatar-image";
-import { useUser } from "@/components/user-provider";
 import { usePoll, useRole } from "@/features/poll/client";
 import { useEditToken } from "@/features/poll/components/mutations";
 import {
@@ -31,6 +30,7 @@ import {
 } from "@/features/poll/components/participant";
 import { useParticipants } from "@/features/poll/components/participants-provider";
 import TruncatedLinkify from "@/features/poll/components/truncated-linkify";
+import { useUser } from "@/features/user/components/user-provider";
 import { Trans, useTranslation } from "@/i18n/client";
 import { RelativeTime } from "@/lib/datetime/relative-time";
 import { requiredString } from "@/lib/utils/form-validation";
